backend/requirements.txt
Flask
git+https://github.com/m-wrzr/populartimes@master
python-dotenv
